requirements.txt
# Core dependencies
streamlit>=1.31.0
langchain>=0.1.10
langchain-openai>=0.0.6
langchain-community>=0.0.25

# PDF processing
pypdf>=4.0.1

# Vector store
faiss-cpu>=1.7.4

# Search tools
tavily-python>=0.3.0
wikipedia>=1.4.0

# Additional utilities
tiktoken>=0.5.2
openai>=1.12.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
